Render "text" as type for datepicker input fields - [x] Refactor the way the datepicker is initialized
This change fixes an issue that appears when you are using a type `date` field. What happens is that google chrome renders its HTML5 input field for dates but complains if the date format is not the same as defined in the users OS settings. You can still save new values, but since the values are localized by Rails they are not being presented in the input field (happens when you enter the page or reload). https://github.com/AlchemyCMS/alchemy_cms/issues/1245