@wip
Stopping now, as I started off by trying to remove the .icon class. Now
I don't think that is the right thing to do. I can do some cleanup and
standardization (adding icon- to the icon classes) without removing the
icon class. I'm going to do that in another branch, in case it turns out
to be wrong.